requirements/dev.txt
-r base.txt

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<7.0
pytest-xdist>=3.5,<4.0
factory-boy>=3.3,<4.0
faker>=25.0,<32.0
vcrpy>=6.0,<7.0
responses>=0.25,<1.0
freezegun>=1.5,<2.0

# Lint / format / type check
ruff>=0.5,<1.0
black>=24.0,<26.0
pre-commit>=3.7,<5.0
mypy>=1.10
django-stubs[compatible-mypy]>=5.0
djangorestframework-stubs[compatible-mypy]>=3.15
types-requests

# Dev tooling
django-debug-toolbar>=4.4,<6.0
django-extensions>=3.2,<4.0
ipython>=8.22
werkzeug>=3.0
